Add config helpers
Refactors config types and adds helper to get config from CLI args

(cherry picked from commit 099e9033da6353a965eadb3de1f9ce94ca7057c7)
