+++ main.go
@@ -7,114 +7,22 @@
 package main
 
 import (
+	"context"
+	"errors"
 	_ "net/http/pprof"
 	"os"
 
-	"github.com/k3s-io/helm-controller/pkg/cmd"
-	"github.com/k3s-io/helm-controller/pkg/version"
+	"github.com/k3s-io/helm-controller/pkg/app"
 	_ "github.com/rancher/wrangler/v3/pkg/generated/controllers/apiextensions.k8s.io"
 	_ "github.com/rancher/wrangler/v3/pkg/generated/controllers/networking.k8s.io"
 	"github.com/rancher/wrangler/v3/pkg/signals"
 	"github.com/sirupsen/logrus"
-	"github.com/urfave/cli/v2"
 )
 
-var config = cmd.HelmController{}
-
 func main() {
-	app := &cli.App{
-		Name:        "helm-controller",
-		Description: "A simple way to manage helm charts with CRDs in K8s.",
-		Version:     version.FriendlyVersion(),
-		Action: func(app *cli.Context) error {
-			return config.Run(app)
-		},
-		Flags: []cli.Flag{
-			&cli.StringFlag{
-				Name:        "controller-name",
-				Value:       "helm-controller",
-				Usage:       "Unique name to identify this controller that is added to all HelmCharts tracked by this controller",
-				EnvVars:     []string{"CONTROLLER_NAME"},
-				Destination: &config.ControllerName,
-			},
-			&cli.BoolFlag{
-				Name:        "debug",
-				Usage:       "Turn on debug logging",
-				Destination: &config.Debug,
-			},
-			&cli.IntFlag{
-				Name:        "debug-level",
-				Usage:       "If debugging is enabled, set klog -v=X",
-				Destination: &config.DebugLevel,
-			},
-			&cli.BoolFlag{
-				Name:        "enforce-pod-limits",
-				Value:       true,
-				Usage:       "Set to false to disable default CPU and memory limits on the pods created to manage helm charts",
-				EnvVars:     []string{"ENFORCE_POD_LIMITS"},
-				Destination: &config.EnforcePodLimits,
-			},
-			&cli.StringFlag{
-				Name:        "default-job-image",
-				Usage:       "Default image to use by jobs managing helm charts",
-				EnvVars:     []string{"DEFAULT_JOB_IMAGE"},
-				Destination: &config.DefaultJobImage,
-			},
-			&cli.StringFlag{
-				Name:        "job-tolerations",
-				Usage:       "JSON array of tolerations to apply to all jobs managing helm charts",
-				EnvVars:     []string{"JOB_TOLERATIONS"},
-				Destination: &config.JobTolerations,
-			},
-			&cli.StringFlag{
-				Name:        "job-cluster-role",
-				Value:       "cluster-admin",
-				Usage:       "Name of the cluster role to use for jobs created to manage helm charts",
-				EnvVars:     []string{"JOB_CLUSTER_ROLE"},
-				Destination: &config.JobClusterRole,
-			},
-			&cli.StringFlag{
-				Name:        "kubeconfig",
-				Usage:       "Kubernetes config files, e.g. $HOME/.kube/config",
-				EnvVars:     []string{"KUBECONFIG"},
-				Destination: &config.Kubeconfig,
-			},
-			&cli.StringFlag{
-				Name:        "master-url",
-				Usage:       "Kubernetes cluster master URL",
-				EnvVars:     []string{"MASTERURL"},
-				Destination: &config.MasterURL,
-			},
-			&cli.StringFlag{
-				Name:        "namespace",
-				Usage:       "Namespace to watch, empty means it will watch CRDs in all namespaces",
-				EnvVars:     []string{"NAMESPACE"},
-				Destination: &config.Namespace,
-			},
-			&cli.StringFlag{
-				Name:        "node-name",
-				Usage:       "Name of the node this controller is running on",
-				EnvVars:     []string{"NODE_NAME"},
-				Destination: &config.NodeName,
-			},
-			&cli.IntFlag{
-				Name:        "pprof-port",
-				Value:       6060,
-				Usage:       "Port to publish HTTP server runtime profiling data in the format expected by the pprof visualization tool. Only enabled if in debug mode",
-				Destination: &config.PprofPort,
-			},
-			&cli.IntFlag{
-				Name:        "threads",
-				Value:       2,
-				Usage:       "Threadiness level to set",
-				EnvVars:     []string{"THREADS"},
-				Destination: &config.Threads,
-			},
-		},
-	}
-
+	app := app.New()
 	ctx := signals.SetupSignalContext()
-	if err := app.RunContext(ctx, os.Args); err != nil {
+	if err := app.RunContext(ctx, os.Args); err != nil && !errors.Is(err, context.Canceled) {
 		logrus.Fatal(err)
 	}
 }
