+++ pkg/cmd/cmd.go
@@ -1,22 +1,20 @@
 package cmd
 
 import (
-	"encoding/json"
+	"context"
 	"fmt"
 	"log"
 	"net/http"
-	"strings"
 	"time"
 
 	"github.com/go-logr/logr"
+	"github.com/k3s-io/helm-controller/pkg/config"
 	"github.com/k3s-io/helm-controller/pkg/controllers"
 	"github.com/k3s-io/helm-controller/pkg/controllers/common"
 	"github.com/k3s-io/helm-controller/pkg/crds"
 	"github.com/rancher/wrangler/v3/pkg/crd"
 	"github.com/rancher/wrangler/v3/pkg/kubeconfig"
 	"github.com/sirupsen/logrus"
-	"github.com/urfave/cli/v2"
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
 	"k8s.io/client-go/tools/clientcmd"
 	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
@@ -28,31 +26,15 @@ const (
 	readyDuration = time.Minute * 1
 )
 
-type HelmController struct {
-	Debug            bool
-	DebugLevel       int
-	Kubeconfig       string
-	MasterURL        string
-	Namespace        string
-	Threads          int
-	ControllerName   string
-	NodeName         string
-	JobClusterRole   string
-	DefaultJobImage  string
-	JobTolerations   string
-	EnforcePodLimits bool
-	PprofPort        int
-}
-
-func (hc *HelmController) SetupLogging() (logr.Logger, error) {
+func SetupLogging(debug bool) (logr.Logger, error) {
 	klog.EnableContextualLogging(true)
-	if hc.Debug {
+	if debug {
 		logrus.SetLevel(logrus.DebugLevel)
 	}
 	return common.NewLogrusSink(nil).AsLogr(), nil
 }
 
-func (hc *HelmController) Run(app *cli.Context) error {
+func Run(ctx context.Context, hc config.CLI) error {
 	if hc.Debug && hc.PprofPort > 0 {
 		go func() {
 			// Serves HTTP server runtime profiling data in the format expected by the
@@ -61,12 +43,13 @@ func (hc *HelmController) Run(app *cli.Context) error {
 			log.Println(http.ListenAndServe(fmt.Sprintf("localhost:%d", hc.PprofPort), nil))
 		}()
 	}
-	logger, err := hc.SetupLogging()
+	logger, err := SetupLogging(hc.Debug)
 	if err != nil {
 		return err
 	}
+	ctx = klog.NewContext(ctx, logger)
 
-	cfg := hc.GetNonInteractiveClientConfig()
+	cfg := getNonInteractiveClientConfig(hc)
 	rest, err := cfg.ClientConfig()
 	if err != nil {
 		return err
@@ -76,28 +59,13 @@ func (hc *HelmController) Run(app *cli.Context) error {
 		return err
 	}
 
-	ctx := klog.NewContext(app.Context, logger)
-
 	crds, err := crds.List()
 	if err != nil {
 		return err
 	}
 
-	tolerations, err := parseTolerations(hc.JobTolerations)
+	opts, err := hc.GetControllerConfig()
 	if err != nil {
-		return fmt.Errorf("invalid --job-tolerations JSON: %w", err)
-	}
-
-	opts := common.Options{
-		Threadiness:      hc.Threads,
-		NodeName:         hc.NodeName,
-		JobClusterRole:   hc.JobClusterRole,
-		DefaultJobImage:  hc.DefaultJobImage,
-		JobTolerations:   tolerations,
-		EnforcePodLimits: hc.EnforcePodLimits,
-	}
-
-	if err := opts.Validate(); err != nil {
 		return err
 	}
 
@@ -110,10 +78,10 @@ func (hc *HelmController) Run(app *cli.Context) error {
 	}
 
 	<-ctx.Done()
-	return nil
+	return ctx.Err()
 }
 
-func (hc *HelmController) GetNonInteractiveClientConfig() clientcmd.ClientConfig {
+func getNonInteractiveClientConfig(hc config.CLI) clientcmd.ClientConfig {
 	// Modified https://github.com/rancher/wrangler/blob/3ecd23dfea3bb4c76cbe8e06fb158eed6ae3dd31/pkg/kubeconfig/loader.go#L12-L32
 	return clientcmd.NewInteractiveDeferredLoadingClientConfig(
 		kubeconfig.GetLoadingRules(hc.Kubeconfig),
@@ -121,18 +89,4 @@ func (hc *HelmController) GetNonInteractiveClientConfig() clientcmd.ClientConfig
 			ClusterDefaults: clientcmd.ClusterDefaults,
 			ClusterInfo:     clientcmdapi.Cluster{Server: hc.MasterURL},
 		}, nil)
-}
-
-// parseTolerations takes the CLI string and parses it into a slice of corev1.Toleration objects
-func parseTolerations(raw string) ([]corev1.Toleration, error) {
-	raw = strings.TrimSpace(raw)
-	if raw == "" {
-		return nil, nil
-	}
-
-	var tolerations []corev1.Toleration
-	if err := json.Unmarshal([]byte(raw), &tolerations); err != nil {
-		return nil, err
-	}
-	return tolerations, nil
 }
